% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exitprob_tsssd}
\alias{exitprob_tsssd}
\title{Exit probabilities for two-stage seamless sequential design (TSSSD)}
\usage{
exitprob_tsssd(
  M = NA_integer_,
  r = 1,
  theta = NA_real_,
  corr_known = TRUE,
  K = NA_integer_,
  b = NA_real_,
  I = NA_real_
)
}
\arguments{
\item{M}{number of active arms in phase 2.}

\item{r}{randomization ratio of each active arm to the common control in phase 2.}

\item{theta}{vector of true treatment effects for the M active arms in phase 2.}

\item{corr_known}{whether the correlation between the Wald statistics in
phase 2 is known. If \code{TRUE}, the correlation is determined by
the randomization ratio \code{r} as \eqn{r / (r + 1)};
if \code{FALSE}, the correlation is conservatively set to 0.}

\item{K}{number of looks in phase 3.}

\item{b}{vector of critical values for phase 2 and the K looks in phase 3.}

\item{I}{vector of information levels for phase 2 and the K looks in
phase 3 for any arm versus the common control.}
}
\value{
A list with three components:
\itemize{
\item exitProb: A vector of length \eqn{K + 1}, where the first element is the
exit probability in phase 2, and the \eqn{k}-th element for
\eqn{k = 1,\ldots,K} is the exit probability at look \eqn{k} in phase 3.
\item exitProbByArm: A matrix of dimension \eqn{(K + 1)} by \eqn{M}, where
the \eqn{(k, m)}-th  element for \eqn{k = 0,\ldots,K} and
\eqn{m = 1,\ldots,M} is the probability of rejecting the
global null hypothesis in phase 2 if arm \eqn{m} is selected in phase 2
when \eqn{k = 0}, and the probability of rejecting the global null hypothesis
at look \eqn{k} in phase 3 if arm \eqn{m} is selected in phase 2 when
\eqn{k = 1,...,K}.
\item selectAsBest: A vector of length \eqn{M}, where the \eqn{m}-th element
is the probability of selecting arm \eqn{m} in phase 2 as the best arm
among the \eqn{M} active arms in phase 2.
}
}
\description{
Computes the exit probabilities for a two-stage selection
and testing design with a common control arm in phase 2 and the selected
active arm and the common control arm in phase 3.
}
\details{
The exit probability in phase 2 is the probability of rejecting
the global null hypothesis at the end of phase 2, and the exit probability
at look \eqn{k} in phase 3 is the probability of rejecting the global null
hypothesis at look \eqn{k} in phase 3 given that the global null hypothesis
is not rejected in phase 2. The exit probabilities are computed under
the multivariate normal distribution of the Wald statistics for the \eqn{M}
active arms in phase 2 and the selected active arm in phase 3, where
the covariance matrix is determined by the randomization ratio \eqn{r}.

We reject the global null hypothesis in phase 2 if \deqn{Z(I_0) >= b_0,}
where \eqn{Z(I_0)} is the Wald statistic for best arm in phase 2, and
we reject the global null hypothesis at look \eqn{k} in phase 3 if
\deqn{Z(I_0) < b_0, Z(I_j) < b_j \text{ for } j = 1,\ldots,k-1,
\text{ and } Z(I_k) >= b_k,} where \eqn{Z(I_k)} is the Wald statistic
for the selected arm at look \eqn{k} in phase 3.

The following assumptions are made for the TSSSD design:
\itemize{
\item Information is the same across active arms in phase 2
\item At most one active arm can be selected at the end of phase 2
\item The selected arm is the one with the largest score statistic in phase 2
}
}
\examples{

# Two active arms vs. a common control in phase 2, and the selected
# active arm vs. the common control in phase 3 with 2 looks in phase 3

# equal spacing of information levels across phase 2/3 for any arm vs.
# the common control

# maximum of 110 patients per arm planned
# standard deviation of the endpoint is 1
I <- c(110 / (2 * 1.0^2) * seq(1, 3)/3)

# O'Brien-Fleming critical values for phase 2/3
b <- c(3.776606, 2.670463, 2.180424)

# Type I error under the global null hypothesis
p0 <- exitprob_tsssd(M = 2, theta = c(0, 0), K = 2, b = b, I = I)
cumsum(p0$exitProb)

# Power under the alternative hypothesis where the treatment effects for
# the two active arms are 0.3 and 0.5, respectively
p1 <- exitprob_tsssd(M = 2, theta = c(0.3, 0.5), K = 2, b = b, I = I)
cumsum(p1$exitProb)

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
