% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{fadjpbon}
\alias{fadjpbon}
\title{Adjusted p-Values for Bonferroni-Based Graphical Approaches}
\usage{
fadjpbon(wgtmat, p)
}
\arguments{
\item{wgtmat}{A list containing the weight matrix and the indicator matrix
for intersection hypotheses.}

\item{p}{The raw p-values for elementary hypotheses.}
}
\value{
A list with the following components:
\itemize{
\item \code{inthyp}: The indicator matrix for the intersection hypotheses.
\item \code{pinter}: The local p-values for the intersection hypotheses.
\item \code{padj}: The adjusted p-values for the elementary hypotheses.
}
}
\description{
Obtains the adjusted p-values for graphical approaches
using weighted Bonferroni tests.
}
\examples{

pvalues <- matrix(c(0.01,0.005,0.015,0.022, 0.02,0.015,0.010,0.023),
                  nrow=2, ncol=4, byrow=TRUE)
w <- c(0.5,0.5,0,0)
G <- matrix(c(0,0,1,0,0,0,0,1,0,1,0,0,1,0,0,0),
            nrow=4, ncol=4, byrow=TRUE)
wgtmat <- fwgtmat(w,G)
fadjpbon(wgtmat, pvalues)

}
\references{
Frank Bretz, Willi Maurer, Werner Brannath and Martin Posch. A
graphical approach to sequentially rejective multiple test
procedures. Statistics in Medicine. 2009; 28:586-604.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
