% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fDefaultWgtmat}
\alias{fDefaultWgtmat}
\title{Default Weight Matrix for All Intersection Hypotheses}
\usage{
fDefaultWgtmat(m)
}
\arguments{
\item{m}{The number of elementary hypotheses.}
}
\value{
A list with the following components:
\itemize{
\item \code{inthyp}: The indicator matrix for the intersection hypotheses.
\item \code{wgtmat}: The default weight matrix for the elementary hypotheses.
}
}
\description{
Obtains the default weight matrix for all intersection
hypotheses, assigning equal weights to the elementary hypotheses within
each intersection hypothesis.
}
\examples{

fDefaultWgtmat(3)

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
