% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_two_stage.R
\name{fNewBound}
\alias{fNewBound}
\title{Compute new stage-wise bounds after adaptation}
\usage{
fNewBound(
  m,
  wgtmat,
  family = NULL,
  corr = NULL,
  stg1_p,
  stg1_inthyp_nr_idx,
  CER,
  stg2_elemhyp_idx,
  stg2_wgtmat,
  info_frac_new,
  nthreads = 0
)
}
\arguments{
\item{m}{Number of hypotheses.}

\item{wgtmat}{Weight matrix for the original intersection hypotheses.}

\item{family}{Family matrix indicating which hypotheses belong to which
families. If NULL, all hypotheses are assumed to belong to the same family.}

\item{corr}{Correlation matrix for the test statistics.}

\item{stg1_p}{P-values for the elementary hypotheses in stage 1.}

\item{stg1_inthyp_nr_idx}{Indices of the intersection hypotheses not rejected
in stage 1.}

\item{CER}{Conditional error rates for the intersection hypotheses.}

\item{stg2_elemhyp_idx}{Indices of the elementary hypotheses in stage 2.}

\item{stg2_wgtmat}{Weight matrix for the intersection hypotheses in stage 2.}

\item{info_frac_new}{New information fraction for stage 1 after adaptation.}

\item{nthreads}{The number of threads to use in simulations (0 means
the default RcppParallel behavior).}
}
\value{
A list containing the adapted intersection hypotheses and their new
stage 2 bounds.
}
\description{
Compute new stage 2 bounds after an adaptation of the
hypotheses tested or their weights.
}
\examples{
initial_weights <- c(0.5, 0.5, 0, 0)
transition_matrix <- matrix(c(0, 0.5, 0.5, 0,
                              0.5, 0, 0, 0.5,
                              0, 1, 0, 0,
                              1, 0, 0, 0),
                            nrow = 4, byrow = TRUE)
wgtmat <- fwgtmat(initial_weights, transition_matrix)
family <- matrix(c(1, 1, 0, 0,
                   0, 0, 1, 1),
                 nrow = 2, byrow = TRUE)
corr <- matrix(c(1, 0.5, NA, NA,
                 0.5, 1, NA, NA,
                 NA, NA, 1, 0.5,
                 NA, NA, 0.5, 1),
               nrow = 4, byrow = TRUE)
stage1_pvalues <- c(0.00045, 0.0952, 0.0225, 0.1104)
bounds <- fStageBound(4, wgtmat, family, corr, 0.025,
                      errorSpent(0.5, 0.025, "sfOF"), 0.5)
conditional_error_rates <- fCER(
  4, wgtmat, family, corr, 0.5,
  bounds$stg1_bnd, bounds$stg2_bnd, stage1_pvalues)
stage2_weight_matrix <- fwgtmat(
  w = c(0.5, 0.5),
  G = matrix(c(0, 1, 1, 0), 2, 2, byrow = TRUE))
fNewBound(4, wgtmat, family, corr, stage1_pvalues,
          conditional_error_rates$stg1_inthyp_nr_idx,
          conditional_error_rates$CER, c(2, 4),
          stage2_weight_matrix, 0.4, nthreads = 1)

}
\references{
Cyrus Mehta, Ajoy Mukhopadhyay, and Martin Posch. Graph Based, Adaptive,
Multiarm, Multiple Endpoint, Two-Stage Designs. Statistics in Medicine.
2025.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
