% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDesignSurvivals.R
\name{fquantile}
\alias{fquantile}
\title{The Quantiles of a Survival Distribution}
\usage{
fquantile(S, probs, ...)
}
\arguments{
\item{S}{The survival function of a univariate survival time.}

\item{probs}{The numeric vector of probabilities.}

\item{...}{Additional arguments to be passed to S.}
}
\value{
A vector of \code{length(probs)} for the quantiles.
}
\description{
Obtains the quantiles of a survival distribution.
}
\examples{

fquantile(pweibull, probs = c(0.25, 0.5, 0.75),
          shape = 1.37, scale = 1/0.818, lower.tail = FALSE)

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
