% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{fstdmix}
\alias{fstdmix}
\title{Adjusted p-Values for Standard Mixture Gatekeeping Procedures}
\usage{
fstdmix(
  p,
  family = NULL,
  serial,
  parallel = NULL,
  gamma,
  test = "hommel",
  exhaust = TRUE
)
}
\arguments{
\item{p}{The raw p-values for elementary hypotheses.}

\item{family}{The matrix of family indicators for the hypotheses.}

\item{serial}{The matrix of serial rejection set for the hypotheses.}

\item{parallel}{The matrix of parallel rejection set for the hypotheses.}

\item{gamma}{The truncation parameters for each family. The truncation
parameter for the last family is automatically set to 1.}

\item{test}{The component multiple testing procedure. Options include
"holm", "hochberg", or "hommel". Defaults to "hommel".}

\item{exhaust}{Whether to use alpha-exhausting component testing procedure
for the last family with active hypotheses. It defaults to \code{TRUE}.}
}
\value{
A list with the following components:
\itemize{
\item \code{inthyp}: The indicator matrix for the intersection hypotheses.
\item \code{pinter}: The local p-values for the intersection hypotheses.
\item \code{padj}: The adjusted p-values for the elementary hypotheses.
}
}
\description{
Obtains the adjusted p-values for the standard gatekeeping
procedures for multiplicity problems involving serial and parallel
logical restrictions.
}
\examples{

p <- c(0.0194, 0.0068, 0.0271, 0.0088, 0.0370, 0.0018, 0.0814, 0.0066)
family <- matrix(c(1, 1, 0, 0, 0, 0, 0, 0,
                   0, 0, 1, 1, 0, 0, 0, 0,
                   0, 0, 0, 0, 1, 1, 0, 0,
                   0, 0, 0, 0, 0, 0, 1, 1),
                 nrow=4, byrow=TRUE)

serial <- matrix(c(0, 0, 0, 0, 0, 0, 0, 0,
                   0, 0, 0, 0, 0, 0, 0, 0,
                   1, 0, 0, 0, 0, 0, 0, 0,
                   0, 1, 0, 0, 0, 0, 0, 0,
                   0, 0, 1, 0, 0, 0, 0, 0,
                   0, 0, 0, 1, 0, 0, 0, 0,
                   0, 0, 0, 0, 1, 0, 0, 0,
                   0, 0, 0, 0, 0, 1, 0, 0),
                 nrow=8, byrow=TRUE)

parallel <- matrix(0, 8, 8)
gamma <- c(0.6, 0.6, 0.6, 1)
fstdmix(p, family, serial, parallel, gamma, test = "hommel",
        exhaust = FALSE)

}
\references{
Alex Dmitrienko and Ajit C Tamhane. Mixtures of multiple testing
procedures for gatekeeping applications in clinical trials.
Statistics in Medicine. 2011; 30(13):1473–1488.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
