% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getADCI_mams}
\alias{getADCI_mams}
\title{Confidence Interval After Adaptation for Multi-Arm Multi-Stage Design}
\usage{
getADCI_mams(
  M = NA_integer_,
  r = NA_real_,
  corr_known = TRUE,
  L = NA_integer_,
  zL = NA_real_,
  IMax = NA_real_,
  kMax = NA_integer_,
  informationRates = NA_real_,
  efficacyStopping = NA_integer_,
  criticalValues = NULL,
  alpha = 0.25,
  typeAlphaSpending = "sfOF",
  parameterAlphaSpending = NA_real_,
  spendingTime = NA_real_,
  MullerSchafer = FALSE,
  MNew = NA_integer_,
  selected = NA_integer_,
  rNew = NA_real_,
  Lc = NA_integer_,
  zLc = NA_real_,
  INew = NA_real_,
  informationRatesNew = NA_real_,
  efficacyStoppingNew = NA_integer_,
  typeAlphaSpendingNew = "sfOF",
  parameterAlphaSpendingNew = NA_real_,
  spendingTimeNew = NA_real_
)
}
\arguments{
\item{M}{Number of active treatment arms in the primary trial.}

\item{r}{Randomization ratio of each active arm to the common control
in the primary trial.}

\item{corr_known}{Logical. If \code{TRUE}, the correlation between Wald
statistics is derived from the randomization ratio \eqn{r}
as \eqn{r / (r + 1)}. If \code{FALSE}, a conservative correlation of
0 is assumed.}

\item{L}{The interim adaptation look of the primary trial.}

\item{zL}{The z-test statistics at the interim adaptation look of
the primary trial.}

\item{IMax}{Maximum information for any active arm versus the common
control for the primary trial. Must be provided.}

\item{kMax}{The maximum number of stages of the primary trial.}

\item{informationRates}{The information rates of the primary trial.}

\item{efficacyStopping}{Indicators of whether efficacy stopping is
allowed at each stage of the primary trial. Defaults to \code{TRUE}
if left unspecified.}

\item{criticalValues}{The matrix of by-level upper boundaries on the
z-test statistic scale for efficacy stopping up to look \code{L}
for the primary trial.
The first column is for level \code{M}, the second column is for
level \code{M - 1}, and so on, with the last column for level 1.
If left unspecified, the critical values will be computed based
on the specified alpha spending function.}

\item{alpha}{The significance level of the primary trial.
Defaults to 0.025.}

\item{typeAlphaSpending}{The type of alpha spending for the primary
trial. One of the following:
\code{"OF"} for O'Brien-Fleming boundaries,
\code{"P"} for Pocock boundaries,
\code{"WT"} for Wang & Tsiatis boundaries,
\code{"sfOF"} for O'Brien-Fleming type spending function,
\code{"sfP"} for Pocock type spending function,
\code{"sfKD"} for Kim & DeMets spending function,
\code{"sfHSD"} for Hwang, Shi & DeCani spending function, and
\code{"none"} for no early efficacy stopping.
Defaults to \code{"sfOF"}.}

\item{parameterAlphaSpending}{The parameter value of alpha spending
for the primary trial. Corresponds to \eqn{\Delta} for "WT",
\eqn{\rho} for "sfKD", and \eqn{\gamma} for "sfHSD".}

\item{spendingTime}{The error spending time of the primary trial.
Defaults to missing, in which case, it is the same as
\code{informationRates}.}

\item{MullerSchafer}{Whether to use the Muller and Schafer (2001) method
for trial adaptation.}

\item{MNew}{The number of active treatment arms in the secondary trial.}

\item{selected}{The indices of the selected treatment arms for the
secondary trial among the \code{M} active arms in the primary trial.}

\item{rNew}{The randomization ratio of each active arm to the common control
in the secondary trial.}

\item{Lc}{The termination look of the integrated trial.}

\item{zLc}{The z-test statistics at the termination look of the
integrated trial.}

\item{INew}{The maximum information for any active arm versus the common
control in the secondary trial.}

\item{informationRatesNew}{The spacing of looks of the secondary trial
up to look \code{L2}.}

\item{efficacyStoppingNew}{The indicators of whether efficacy stopping is
allowed at each look of the secondary trial up to look \code{L2}.
Defaults to \code{TRUE} if left unspecified.}

\item{typeAlphaSpendingNew}{The type of alpha spending for the secondary
trial. One of the following:
\code{"OF"} for O'Brien-Fleming boundaries,
\code{"sfOF"} for O'Brien-Fleming type spending function,
\code{"sfP"} for Pocock type spending function,
\code{"sfKD"} for Kim & DeMets spending function,
\code{"sfHSD"} for Hwang, Shi & DeCani spending function, and
\code{"none"} for no early efficacy stopping.
Defaults to \code{"sfOF"}.}

\item{parameterAlphaSpendingNew}{The parameter value of alpha spending
for the secondary trial. Corresponds to \eqn{\Delta} for "WT",
\eqn{\rho} for "sfKD", and \eqn{\gamma} for "sfHSD".}

\item{spendingTimeNew}{The error spending time of the secondary trial
up to look \code{L2}. Defaults to missing, in which case, it is
the same as \code{informationRatesNew}.}
}
\value{
A data frame with the following variables:
\itemize{
\item \code{level}: Number of individual hypotheses considered for multiplicity.
\item \code{index}: The index of the treatment arm among the \code{M} active arms.
\item \code{pvalue}: p-value for rejecting the null hypothesis.
\item \code{thetahat}: Median unbiased point estimate of the parameter.
\item \code{cilevel}: Confidence interval level.
\item \code{lower}: Lower bound of confidence interval.
\item \code{upper}: Upper bound of confidence interval.
}
}
\description{
Obtains the p-value, conservative point estimate, and
confidence interval after the end of an adaptive multi-arm multi-stage trial.
}
\examples{
getADCI_mams(
  M = 2, r = 1, corr_known = FALSE, L = 1, zL = c(2.075, 2.264),
  IMax = 300 / 4, kMax = 2, informationRates = c(0.5, 1),
  alpha = 0.025, typeAlphaSpending = "sfOF",
  MNew = 1, selected = 2, rNew = 1,
  Lc = 2, zLc = 1.667, INew = 374 / 4)

}
\references{
Ping Gao, Yingqiu Li.
Adaptive multiple comparison sequential design (AMCSD) for clinical trials.
Journal of Biopharmaceutical Statistics, 2024, 34(3), 424-440.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
