% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBound}
\alias{getBound}
\title{Efficacy Boundaries for Group Sequential Design}
\usage{
getBound(
  k = NA_integer_,
  informationRates = NA_real_,
  alpha = 0.025,
  typeAlphaSpending = "sfOF",
  parameterAlphaSpending = NA_real_,
  userAlphaSpending = NA_real_,
  spendingTime = NA_real_,
  efficacyStopping = NA_integer_
)
}
\arguments{
\item{k}{Look number for the current analysis.}

\item{informationRates}{Information rates up to the current look. Must be
increasing and less than or equal to 1.}

\item{alpha}{The significance level. Defaults to 0.025.}

\item{typeAlphaSpending}{The type of alpha spending. One of the following:
\code{"OF"} for O'Brien-Fleming boundaries,
\code{"P"} for Pocock boundaries,
\code{"WT"} for Wang & Tsiatis boundaries,
\code{"sfOF"} for O'Brien-Fleming type spending function,
\code{"sfP"} for Pocock type spending function,
\code{"sfKD"} for Kim & DeMets spending function,
\code{"sfHSD"} for Hwang, Shi & DeCani spending function,
\code{"user"} for user defined spending, and
\code{"none"} for no early efficacy stopping.
Defaults to \code{"sfOF"}.}

\item{parameterAlphaSpending}{The parameter value for the alpha spending.
Corresponds to \eqn{\Delta} for \code{"WT"}, \eqn{\rho} for \code{"sfKD"},
and \eqn{\gamma} for \code{"sfHSD"}.}

\item{userAlphaSpending}{The user defined alpha spending. Cumulative alpha
spent up to each stage.}

\item{spendingTime}{A vector of length \code{k} for the error spending
time at each analysis. Must be increasing and less than or equal to 1.
Defaults to missing, in which case, it is the same as
\code{informationRates}.}

\item{efficacyStopping}{Indicators of whether efficacy stopping is allowed
at each stage. Defaults to \code{TRUE} if left unspecified.}
}
\value{
A numeric vector of critical values up to the current look.
}
\description{
Obtains the efficacy stopping boundaries for a group
sequential design.
}
\details{
If \code{typeAlphaSpending} is "OF", "P", or "WT", then the boundaries
will be based on equally spaced looks.
}
\examples{

getBound(k = 2, informationRates = c(0.5,1),
         alpha = 0.025, typeAlphaSpending = "sfOF")

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
