% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getDesign_tsssd}
\alias{getDesign_tsssd}
\title{Power and Sample Size for a Generic Two-Stage Seamless Sequential Design}
\usage{
getDesign_tsssd(
  beta = NA_real_,
  IMax = NA_real_,
  theta = NA_real_,
  M = NA_integer_,
  r = 1,
  corr_known = TRUE,
  K = 1L,
  informationRates = NA_real_,
  efficacyStopping = NA_integer_,
  criticalValues = NA_real_,
  alpha = 0.025,
  typeAlphaSpending = "sfOF",
  parameterAlphaSpending = NA_real_,
  userAlphaSpending = NA_real_,
  spendingTime = NA_real_,
  varianceRatio = 1
)
}
\arguments{
\item{beta}{The type II error.}

\item{IMax}{The maximum information for any active arm versus the common
control. Either \code{beta} or \code{IMax} should be provided while
the other one should be missing.}

\item{theta}{The parameter vector. The global null hypothesis is at
\eqn{\theta_i = 0} for all \eqn{i}, and the alternative hypothesis
is one-sided for \eqn{theta_i > 0} for at least one \eqn{i = 1,\ldots,M}.}

\item{M}{The number of active arms in phase 2. Must be at least 2.}

\item{r}{The randomization ratio of each active arm to the common control
in phase 2. Must be positive.}

\item{corr_known}{whether the correlation between the Wald statistics in
phase 2 is known. If \code{TRUE}, the correlation is determined by
the randomization ratio \code{r} as \eqn{r / (r + 1)};
if \code{FALSE}, the correlation is conservatively set to 0.}

\item{K}{The maximum number of stages in phase 3. Must be at least 1.}

\item{informationRates}{The information rates. Fixed prior to the trial.
Defaults to \code{(1:(K+1)) / (K+1)} if left unspecified.}

\item{efficacyStopping}{Indicators of whether efficacy stopping is allowed
at each stage. Defaults to true if left unspecified.}

\item{criticalValues}{Upper boundaries on the z-test statistic scale
for stopping for efficacy.}

\item{alpha}{The significance level. Defaults to 0.025.}

\item{typeAlphaSpending}{The type of alpha spending. One of the following:
"OF" for O'Brien-Fleming boundaries, "P" for Pocock boundaries,
"WT" for Wang & Tsiatis boundaries, "sfOF" for O'Brien-Fleming type
spending function, "sfP" for Pocock type spending function,
"sfKD" for Kim & DeMets spending function, "sfHSD" for Hwang,
Shi & DeCani spending function, "user" for user defined spending,
and "none" for no early efficacy stopping. Defaults to "sfOF".}

\item{parameterAlphaSpending}{The parameter value for the alpha spending.
Corresponds to Delta for "WT", rho for "sfKD", and gamma for "sfHSD".}

\item{userAlphaSpending}{The user defined alpha spending. Cumulative alpha
spent up to each stage.}

\item{spendingTime}{A vector of length \code{K+1} for the error spending
time at each analysis. Defaults to missing, in which case, it is the
same as \code{informationRates}.}

\item{varianceRatio}{The ratio of the variance under H0 to the
variance under H1.}
}
\value{
An S3 class \code{design} object with three components:
\itemize{
\item \code{overallResults}: A data frame containing the following variables:
\itemize{
\item \code{overallReject}: The overall rejection probability.
\item \code{alpha}: The overall significance level.
\item \code{M}: The number of active arms in phase 2.
\item \code{r}: The randomization ratio of each active arm to the
common control in phase 2.
\item \code{corr_known}: Whether the correlation between the Wald statistics
in phase 2 is known.
\item \code{K}: The number of stages in phase 3.
\item \code{information}: The maximum information for any given active arm
versus the common control.
}
\item \code{byStageResults}: A data frame containing the following variables:
\itemize{
\item \code{informationRates}: The information rates.
\item \code{efficacyBounds}: The efficacy boundaries on the Z-scale.
\item \code{rejectPerStage}: The probability for efficacy stopping.
\item \code{cumulativeRejection}: The cumulative probability for efficacy
stopping.
\item \code{cumulativeAlphaSpent}: The cumulative alpha spent.
\item \code{efficacyTheta}: The efficacy boundaries on the parameter
scale.
\item \code{efficacyP}: The efficacy boundaries on the p-value scale.
\item \code{information}: The cumulative information.
\item \code{efficacyStopping}: Whether to allow efficacy stopping.
}
\item \code{byArmResults}: A data frame containing the following variables:
\itemize{
\item \code{theta}: The parameter values for the active arms.
\item \code{selectAsBest}: The probability of being selected as the
best arm at the end of phase 2.
\item \code{powerByArm}: The probability of rejecting the null hypothesis
for each arm at the end of the trial.
\item \code{condPowerByArm}: The conditional power for each arm given that
the arm is selected as the best at the end of phase 2.
}
\item \code{settings}: A list containing the following input parameters:
\itemize{
\item \code{typeAlphaSpending}: The type of alpha spending.
\item \code{parameterAlphaSpending}: The parameter value for alpha
spending.
\item \code{userAlphaSpending}: The user defined alpha spending.
\item \code{spendingTime}: The error spending time at each analysis.
\item \code{varianceRatio}: The ratio of the variance under H0
to the variance under H1.
}
}
}
\description{
Obtains the maximum information and stopping boundaries
for a generic two-stage seamless sequential design design, or obtains
the power given the maximum information and stopping boundaries.
}
\details{
When \code{corr_known} is \code{FALSE}, the critical boundaries
are computed under the assumption of independence among the Wald
statistics in phase 2, whereas the power is evaluated using the
correlation implied by the randomization ratio.
}
\examples{

# Example 1: obtain the maximum information given power
(design1 <- getDesign_tsssd(
  beta = 0.1, theta = c(0.3, 0.5), M = 2, r = 1.0,
  K = 2, informationRates = seq(1, 3)/3,
  alpha = 0.025, typeAlphaSpending = "OF"))

# Example 2: obtain power given the maximum information
(design2 <- getDesign_tsssd(
  IMax = 110/(2*1^2), theta = c(0.3, 0.5), M = 2, r = 1.0,
  K = 2, informationRates = seq(1, 3)/3,
  alpha = 0.025, typeAlphaSpending = "OF"))

}
\references{
Ping Gao, Yingqiu Li.
Adaptive two-stage seamless sequential design for clinical trials.
Journal of Biopharmaceutical Statistics, 2025, 35(4), 565-587.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
