% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrsim.R
\name{lrsim2e}
\alias{lrsim2e}
\title{Log-Rank Test Simulation for PFS and OS Endpoints}
\usage{
lrsim2e(
  kMax = 1,
  kMaxpfs = 1,
  hazardRatioH0pfs = 1,
  hazardRatioH0os = 1,
  allocation1 = 1,
  allocation2 = 1,
  accrualTime = 0,
  accrualIntensity = NA,
  piecewiseSurvivalTime = 0,
  stratumFraction = 1,
  rho_pd_os = 0,
  lambda1pfs = NA,
  lambda2pfs = NA,
  lambda1os = NA,
  lambda2os = NA,
  gamma1pfs = 0,
  gamma2pfs = 0,
  gamma1os = 0,
  gamma2os = 0,
  n = NA,
  followupTime = NA,
  fixedFollowup = FALSE,
  rho1 = 0,
  rho2 = 0,
  plannedEvents = NA,
  plannedTime = NA,
  maxNumberOfIterations = 1000,
  maxNumberOfRawDatasetsPerStage = 0,
  seed = 0,
  nthreads = 0
)
}
\arguments{
\item{kMax}{The maximum number of stages.}

\item{kMaxpfs}{Number of stages with timing determined by PFS events.
Ranges from 0 (none) to \code{kMax}.}

\item{hazardRatioH0pfs}{Hazard ratio under the null hypothesis for the
active treatment vs control for PFS. Defaults to 1 for
superiority test.}

\item{hazardRatioH0os}{Hazard ratio under the null hypothesis for the
active treatment vs control for OS. Defaults to 1 for
superiority test.}

\item{allocation1}{Number of subjects in the treatment group in
a randomization block. Defaults to 1 for equal randomization.}

\item{allocation2}{Number of subjects in the control group in
a randomization block. Defaults to 1 for equal randomization.}

\item{accrualTime}{A vector that specifies the starting time of
piecewise Poisson enrollment time intervals. Must start with 0, e.g.,
\code{c(0, 3)} breaks the time axis into 2 accrual intervals:
\eqn{[0, 3)} and \eqn{[3, \infty)}.}

\item{accrualIntensity}{A vector of accrual intensities. One for
each accrual time interval.}

\item{piecewiseSurvivalTime}{A vector that specifies the starting time of
piecewise exponential survival time intervals. Must start with 0, e.g.,
\code{c(0, 6)} breaks the time axis into 2 event intervals:
\eqn{[0, 6)} and \eqn{[6, \infty)}.
Defaults to 0 for exponential distribution.}

\item{stratumFraction}{A vector of stratum fractions that sum to 1.
Defaults to 1 for no stratification.}

\item{rho_pd_os}{The correlation coefficient for the standard
bivariate normal random variables used to generate time to disease
progression (PD) and time to death using the inverse CDF method.}

\item{lambda1pfs}{A vector of hazard rates for the event in each analysis
time interval by stratum for the treatment group and PFS.}

\item{lambda2pfs}{A vector of hazard rates for the event in each analysis
time interval by stratum for the control group and PFS.}

\item{lambda1os}{A vector of hazard rates for the event in each analysis
time interval by stratum for the treatment group and OS.}

\item{lambda2os}{A vector of hazard rates for the event in each analysis
time interval by stratum for the control group and OS.}

\item{gamma1pfs}{The hazard rate for exponential dropout, a vector of
hazard rates for piecewise exponential dropout applicable for all
strata, or a vector of hazard rates for dropout in each analysis time
interval by stratum for the treatment group and PFS.}

\item{gamma2pfs}{The hazard rate for exponential dropout, a vector of
hazard rates for piecewise exponential dropout applicable for all
strata, or a vector of hazard rates for dropout in each analysis time
interval by stratum for the control group and PFS.}

\item{gamma1os}{The hazard rate for exponential dropout, a vector of
hazard rates for piecewise exponential dropout applicable for all
strata, or a vector of hazard rates for dropout in each analysis time
interval by stratum for the treatment group and OS.}

\item{gamma2os}{The hazard rate for exponential dropout, a vector of
hazard rates for piecewise exponential dropout applicable for all
strata, or a vector of hazard rates for dropout in each analysis time
interval by stratum for the control group and OS.}

\item{n}{Sample size.}

\item{followupTime}{Follow-up time for the last enrolled subject.}

\item{fixedFollowup}{Whether a fixed follow-up design is used.
Defaults to \code{FALSE} for variable follow-up.}

\item{rho1}{The first parameter of the Fleming-Harrington family of
weighted log-rank test. Defaults to 0 for conventional log-rank test.}

\item{rho2}{The second parameter of the Fleming-Harrington family of
weighted log-rank test. Defaults to 0 for conventional log-rank test.}

\item{plannedEvents}{The planned cumulative total number of PFS events at
Look 1 to Look \code{kMaxpfs} and the planned cumulative total number
of OS events at Look \code{kMaxpfs+1} to Look \code{kMax}.}

\item{plannedTime}{The calendar times for the analyses. To use calendar
time to plan the analyses, \code{plannedEvents} should be missing.}

\item{maxNumberOfIterations}{The number of simulation iterations.
Defaults to 1000.}

\item{maxNumberOfRawDatasetsPerStage}{The number of raw datasets per
stage to extract.}

\item{seed}{The seed to reproduce the simulation results.}

\item{nthreads}{The number of threads to use in simulations (0 means
the default RcppParallel behavior).}
}
\value{
A list with 2 components:
\itemize{
\item \code{sumdata}: A data frame of summary data by iteration and stage:
\itemize{
\item \code{iterationNumber}: The iteration number.
\item \code{eventsNotAchieved}: Whether the target number of events
is not achieved for the iteration.
\item \code{stageNumber}: The stage number, covering all stages even if
the trial stops at an interim look.
\item \code{analysisTime}: The time for the stage since trial start.
\item \code{accruals1}: The number of subjects enrolled at the stage for
the treatment group.
\item \code{accruals2}: The number of subjects enrolled at the stage for
the control group.
\item \code{totalAccruals}: The total number of subjects enrolled at
the stage.
\item \code{endpoint}: The endpoint (1 for PFS or 2 for OS) under
consideration.
\item \code{events1}: The number of events at the stage for
the treatment group.
\item \code{events2}: The number of events at the stage for
the control group.
\item \code{totalEvents}: The total number of events at the stage.
\item \code{dropouts1}: The number of dropouts at the stage for
the treatment group.
\item \code{dropouts2}: The number of dropouts at the stage for
the control group.
\item \code{totalDropouts}: The total number of dropouts at the stage.
\item \code{uscore}: The numerator of the log-rank test statistic for
the endpoint.
\item \code{vscore}: The variance of the log-rank test statistic for
the endpoint.
\item \code{logRankStatistic}: The log-rank test Z-statistic for
the endpoint.
}
\item \code{rawdata} (exists if \code{maxNumberOfRawDatasetsPerStage} is a
positive integer): A data frame for subject-level data for selected
replications, containing the following variables:
\itemize{
\item \code{iterationNumber}: The iteration number.
\item \code{stageNumber}: The stage under consideration.
\item \code{analysisTime}: The time for the stage since trial start.
\item \code{subjectId}: The subject ID.
\item \code{arrivalTime}: The enrollment time for the subject.
\item \code{stratum}: The stratum for the subject.
\item \code{treatmentGroup}: The treatment group (1 or 2) for the
subject.
\item \code{endpoint}: The endpoint (1 for PFS or 2 for OS) under
consideration for the row. Each subject will have two rows, one
for each endpoint.
\item \code{survivalTime}: The underlying survival time for the
event endpoint for the subject.
\item \code{dropoutTime}: The underlying dropout time for the
event endpoint for the subject.
\item \code{timeUnderObservation}: The time under observation
since randomization for the event endpoint for the subject.
\item \code{event}: Whether the subject experienced the event
endpoint.
\item \code{dropoutEvent}: Whether the subject dropped out for the
endpoint.
}
}
}
\description{
Performs simulation for two-endpoint (PFS and OS) two-arm
group sequential trials based on weighted log-rank test. The first
\code{kMaxpfs} looks are driven by the total number of PFS events in
two arms combined, and the subsequent looks are driven by the total
number of OS events in two arms combined. Alternatively,
the analyses can be planned to occur at specified calendar times.
}
\examples{

sim1 = lrsim2e(
  kMax = 3,
  kMaxpfs = 2,
  allocation1 = 2,
  allocation2 = 1,
  accrualTime = c(0, 8),
  accrualIntensity = c(10, 28),
  piecewiseSurvivalTime = 0,
  rho_pd_os = 0,
  lambda1pfs = log(2)/12*0.60,
  lambda2pfs = log(2)/12,
  lambda1os = log(2)/30*0.65,
  lambda2os = log(2)/30,
  n = 420,
  plannedEvents = c(186, 259, 183),
  maxNumberOfIterations = 1000,
  maxNumberOfRawDatasetsPerStage = 1,
  seed = 314159,
  nthreads = 1)

head(sim1$sumdata)
head(sim1$rawdata)

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
