% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_accrualDuration}
\alias{param_accrualDuration}
\title{Parameter Description: accrualDuration}
\arguments{
\item{accrualDuration}{Duration of the enrollment period.}
}
\description{
Parameter Description: accrualDuration
}
\keyword{internal}
