% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_accrualIntensity}
\alias{param_accrualIntensity}
\title{Parameter Description: accrualIntensity}
\arguments{
\item{accrualIntensity}{A vector of accrual intensities. One for
each accrual time interval.}
}
\description{
Parameter Description: accrualIntensity
}
\keyword{internal}
