% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_accrualTime}
\alias{param_accrualTime}
\title{Parameter Description: accrualTime}
\arguments{
\item{accrualTime}{A vector that specifies the starting time of
piecewise Poisson enrollment time intervals. Must start with 0, e.g.,
\code{c(0, 3)} breaks the time axis into 2 accrual intervals:
[0, 3) and [3, Inf).}
}
\description{
Parameter Description: accrualTime
}
\keyword{internal}
