% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_allocationRatioPlanned}
\alias{param_allocationRatioPlanned}
\title{Parameter Description: allocationRatioPlanned}
\arguments{
\item{allocationRatioPlanned}{Allocation ratio for the active treatment
versus control. Defaults to 1 for equal randomization.}
}
\description{
Parameter Description: allocationRatioPlanned
}
\keyword{internal}
