% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_alpha}
\alias{param_alpha}
\title{Parameter Description: alpha}
\arguments{
\item{alpha}{The significance level. Defaults to 0.025.}
}
\description{
Parameter Description: alpha
}
\keyword{internal}
