% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_criticalValues}
\alias{param_criticalValues}
\title{Parameter Description: criticalValues}
\arguments{
\item{criticalValues}{Upper boundaries on the z-test statistic scale
for stopping for efficacy.}
}
\description{
Parameter Description: criticalValues
}
\keyword{internal}
