% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_efficacyStopping}
\alias{param_efficacyStopping}
\title{Parameter Description: efficacyStopping}
\arguments{
\item{efficacyStopping}{Indicators of whether efficacy stopping is allowed
at each stage. Defaults to true if left unspecified.}
}
\description{
Parameter Description: efficacyStopping
}
\keyword{internal}
