% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_estimateHazardRatio}
\alias{param_estimateHazardRatio}
\title{Parameter Description: estimateHazardRatio}
\arguments{
\item{estimateHazardRatio}{Whether to estimate the hazard ratio from
weighted Cox regression model and report the stopping boundaries on
the hazard ratio scale.}
}
\description{
Parameter Description: estimateHazardRatio
}
\keyword{internal}
