% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_fixedFollowup}
\alias{param_fixedFollowup}
\title{Parameter Description: fixedFollowup}
\arguments{
\item{fixedFollowup}{Whether a fixed follow-up design is used.
Defaults to 0 for variable follow-up.}
}
\description{
Parameter Description: fixedFollowup
}
\keyword{internal}
