% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_followupTime}
\alias{param_followupTime}
\title{Parameter Description: followupTime}
\arguments{
\item{followupTime}{Follow-up time for the last enrolled subject.}
}
\description{
Parameter Description: followupTime
}
\keyword{internal}
