% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_futilityStopping}
\alias{param_futilityStopping}
\title{Parameter Description: futilityStopping}
\arguments{
\item{futilityStopping}{Indicators of whether futility stopping is allowed
at each stage. Defaults to true if left unspecified.}
}
\description{
Parameter Description: futilityStopping
}
\keyword{internal}
