% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_gamma}
\alias{param_gamma}
\title{Parameter Description: gamma}
\arguments{
\item{gamma}{The hazard rate for exponential dropout, or a vector of
hazard rates for piecewise exponential dropout. Defaults to 0 for
no dropout.}
}
\description{
Parameter Description: gamma
}
\keyword{internal}
