% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_gamma1}
\alias{param_gamma1}
\title{Parameter Description: gamma1}
\arguments{
\item{gamma1}{The hazard rate for exponential dropout, or a vector of
hazard rates for piecewise exponential dropout for the active
treatment group. Defaults to 0 for no dropout.}
}
\description{
Parameter Description: gamma1
}
\keyword{internal}
