% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_gamma1_stratified}
\alias{param_gamma1_stratified}
\title{Parameter Description: gamma1_stratified}
\arguments{
\item{gamma1}{The hazard rate for exponential dropout, a vector of hazard
rates for piecewise exponential dropout applicable for all strata, or a
vector of hazard rates for dropout in each analysis time interval
by stratum for the active treatment group.}
}
\description{
Parameter Description: gamma1_stratified
}
\keyword{internal}
