% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_gamma2}
\alias{param_gamma2}
\title{Parameter Description: gamma2}
\arguments{
\item{gamma2}{The hazard rate for exponential dropout, or a vector of
hazard rates for piecewise exponential dropout for the control group.
Defaults to 0 for no dropout.}
}
\description{
Parameter Description: gamma2
}
\keyword{internal}
