% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_hazardRatioH0}
\alias{param_hazardRatioH0}
\title{Parameter Description: hazardRatioH0}
\arguments{
\item{hazardRatioH0}{Hazard ratio under the null hypothesis for the active
treatment versus control. Defaults to 1 for superiority test.}
}
\description{
Parameter Description: hazardRatioH0
}
\keyword{internal}
