% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_informationRates}
\alias{param_informationRates}
\title{Parameter Description: informationRates}
\arguments{
\item{informationRates}{The information rates in terms of number of
events. Fixed prior to the trial. Defaults to \code{(1:kMax) / kMax}
if left unspecified.}
}
\description{
Parameter Description: informationRates
}
\keyword{internal}
