% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_lambda}
\alias{param_lambda}
\title{Parameter Description: lambda}
\arguments{
\item{lambda}{A vector of hazard rates for the event. One for
each analysis time interval.}
}
\description{
Parameter Description: lambda
}
\keyword{internal}
