% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_lambda1}
\alias{param_lambda1}
\title{Parameter Description: lambda1}
\arguments{
\item{lambda1}{A vector of hazard rates for the event for the
active treatment group. One for each analysis time interval.}
}
\description{
Parameter Description: lambda1
}
\keyword{internal}
