% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_lambda2}
\alias{param_lambda2}
\title{Parameter Description: lambda2}
\arguments{
\item{lambda2}{A vector of hazard rates for the event for the
control group. One for each analysis time interval.}
}
\description{
Parameter Description: lambda2
}
\keyword{internal}
