% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_lambda2_stratified}
\alias{param_lambda2_stratified}
\title{Parameter Description: lambda2_stratified}
\arguments{
\item{lambda2}{A vector of hazard rates for the event in each analysis
time interval by stratum for the control group.}
}
\description{
Parameter Description: lambda2_stratified
}
\keyword{internal}
