% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_maxFollowupTime}
\alias{param_maxFollowupTime}
\title{Parameter Description: maxFollowupTime}
\arguments{
\item{maxFollowupTime}{Follow-up time for the first enrolled subject.
For fixed followup, \code{maxFollowupTime = minFollowupTime}.
For variable followup,
\code{maxFollowupTime = accrualDuration + minFollowupTime}.}
}
\description{
Parameter Description: maxFollowupTime
}
\keyword{internal}
