% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_minFollowupTime}
\alias{param_minFollowupTime}
\title{Parameter Description: minFollowupTime}
\arguments{
\item{minFollowupTime}{Follow-up time for the last enrolled subject.}
}
\description{
Parameter Description: minFollowupTime
}
\keyword{internal}
