% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_numSubintervals}
\alias{param_numSubintervals}
\title{Parameter Description: numSubintervals}
\arguments{
\item{numSubintervals}{Number of sub-intervals to approximate the mean
and variance of the weighted log-rank test score statistic.
Defaults to 300. Specify a larger number for better approximation.}
}
\description{
Parameter Description: numSubintervals
}
\keyword{internal}
