% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_parameterBetaSpending}
\alias{param_parameterBetaSpending}
\title{Parameter Description: parameterBetaSpending}
\arguments{
\item{parameterBetaSpending}{The parameter value for the beta spending.
Corresponds to \eqn{\rho} for \code{"sfKD"}, and
\eqn{\gamma} for \code{"sfHSD"}.}
}
\description{
Parameter Description: parameterBetaSpending
}
\keyword{internal}
