% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_piecewiseSurvivalTime}
\alias{param_piecewiseSurvivalTime}
\title{Parameter Description: piecewiseSurvivalTime}
\arguments{
\item{piecewiseSurvivalTime}{A vector that specifies the starting time of
piecewise exponential survival time intervals. Must start with 0, e.g.,
\code{c(0, 6)} breaks the time axis into 2 event intervals:
[0, 6) and [6, Inf).
Defaults to 0 for exponential distribution.}
}
\description{
Parameter Description: piecewiseSurvivalTime
}
\keyword{internal}
