% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_rho2}
\alias{param_rho2}
\title{Parameter Description: rho2}
\arguments{
\item{rho2}{The second parameter of the Fleming-Harrington family of
weighted log-rank test. Defaults to 0 for conventional log-rank test.}
}
\description{
Parameter Description: rho2
}
\keyword{internal}
