% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_stratumFraction}
\alias{param_stratumFraction}
\title{Parameter Description: stratumFraction}
\arguments{
\item{stratumFraction}{A vector of stratum fractions that sum to 1.
Defaults to 1 for no stratification.}
}
\description{
Parameter Description: stratumFraction
}
\keyword{internal}
