% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_typeOfComputation}
\alias{param_typeOfComputation}
\title{Parameter Description: typeOfComputation}
\arguments{
\item{typeOfComputation}{The type of computation,
either \code{"direct"} for the direct approximation method,
or \code{"schoenfeld"} for the Schoenfeld method.
Defaults to empty, which selects the Schoenfeld method
under proportional hazards and ordinary log-rank test
and the direct method otherwise.}
}
\description{
Parameter Description: typeOfComputation
}
\keyword{internal}
