% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_userAlphaSpending}
\alias{param_userAlphaSpending}
\title{Parameter Description: userAlphaSpending}
\arguments{
\item{userAlphaSpending}{The user defined alpha spending. Cumulative alpha
spent up to each stage.}
}
\description{
Parameter Description: userAlphaSpending
}
\keyword{internal}
