% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_userBetaSpending}
\alias{param_userBetaSpending}
\title{Parameter Description: userBetaSpending}
\arguments{
\item{userBetaSpending}{The user defined beta spending. Cumulative beta
spent up to each stage.}
}
\description{
Parameter Description: userBetaSpending
}
\keyword{internal}
