% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.adaptDesign}
\alias{print.adaptDesign}
\title{Print adaptive group sequential design}
\usage{
\method{print}{adaptDesign}(x, ...)
}
\arguments{
\item{x}{The adaptDesign object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the primary and second trial information for
an adaptive group sequential design.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
