% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.BOINTable}
\alias{print.BOINTable}
\title{Print the decision table for the BOIN design}
\usage{
\method{print}{BOINTable}(x, ...)
}
\arguments{
\item{x}{The BOINTable object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the decision table of the Bayesian optimal interval
design for MTD finding.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
