% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.design}
\alias{print.design}
\title{Print group sequential design}
\usage{
\method{print}{design}(x, ...)
}
\arguments{
\item{x}{The design object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the stopping boundaries and information inflation
factor for group sequential design.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
