% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designAgreement}
\alias{print.designAgreement}
\title{Print the power and sample size for Cohen's kappa.}
\usage{
\method{print}{designAgreement}(x, ...)
}
\arguments{
\item{x}{The designAgreement object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the power and sample size for Cohen's kappa.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
