% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designANOVA}
\alias{print.designANOVA}
\title{Print the power and sample size for one-way ANOVA}
\usage{
\method{print}{designANOVA}(x, ...)
}
\arguments{
\item{x}{The designANOVA object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the power and sample size for one-way analysis
of variance.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
