% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designANOVAContrast}
\alias{print.designANOVAContrast}
\title{Print the power and sample size for a single contrast in
one-way ANOVA}
\usage{
\method{print}{designANOVAContrast}(x, ...)
}
\arguments{
\item{x}{The designANOVAContrast object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the power and sample size for a single contrast in
one-way analysis of variance.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
