% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designEquiv}
\alias{print.designEquiv}
\title{Print group sequential equivalence design}
\usage{
\method{print}{designEquiv}(x, ...)
}
\arguments{
\item{x}{The design object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the stopping boundaries for group sequential
equivalence design.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
