% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designLogistic}
\alias{print.designLogistic}
\title{Print the power and sample size for logistic regression}
\usage{
\method{print}{designLogistic}(x, ...)
}
\arguments{
\item{x}{The designLogistic object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the power and sample size for logistic regression.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
