% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designMeanDiff}
\alias{print.designMeanDiff}
\title{Print power and sample size results for two-sample mean difference}
\usage{
\method{print}{designMeanDiff}(x, ...)
}
\arguments{
\item{x}{The designMeanDiff object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
two-sample mean difference.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
