% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designMeanDiffCarryoverEquiv}
\alias{print.designMeanDiffCarryoverEquiv}
\title{Print Power and Sample Size Results for Equivalence in Direct
Treatment Effects in Crossover Trials With or Without Accounting for
Carryover Effects}
\usage{
\method{print}{designMeanDiffCarryoverEquiv}(x, ...)
}
\arguments{
\item{x}{The designMeanDiffCarryoverEquiv object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
equivalence in direct treatment effects in crossover trials with or
without accounting for carryover effects.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
