% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designMeanDiffXO}
\alias{print.designMeanDiffXO}
\title{Print power and sample size results for mean difference in 2x2
crossover}
\usage{
\method{print}{designMeanDiffXO}(x, ...)
}
\arguments{
\item{x}{The designMeanDiffXO object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
mean difference in 2x2 crossover.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
