% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designMeanRatio}
\alias{print.designMeanRatio}
\title{Print power and sample size results for two-sample mean ratio}
\usage{
\method{print}{designMeanRatio}(x, ...)
}
\arguments{
\item{x}{The designMeanRatio object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
two-sample mean ratio.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
