% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designMeanRatioXOEquiv}
\alias{print.designMeanRatioXOEquiv}
\title{Print power and sample size results for equivalence in
mean ratio in 2x2 crossover}
\usage{
\method{print}{designMeanRatioXOEquiv}(x, ...)
}
\arguments{
\item{x}{The designMeanRatioXOEquiv object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
equivalence in mean ratio in 2x2 crossover.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
