% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designOneProportion}
\alias{print.designOneProportion}
\title{Print power and sample size results for one-sample proportion}
\usage{
\method{print}{designOneProportion}(x, ...)
}
\arguments{
\item{x}{The designOneProportion object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
one-sample proportion.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
