% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designOrderedBinom}
\alias{print.designOrderedBinom}
\title{Print Power and Sample Size Results for Cochran-Armitage Trend Test
for Ordered Multi-Sample Binomial Response}
\usage{
\method{print}{designOrderedBinom}(x, ...)
}
\arguments{
\item{x}{The designOrderedBinom object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the power and sample size for Cochran-Armitage
trend test for ordered multi-sample binomial response.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
