% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designPairedMeanDiff}
\alias{print.designPairedMeanDiff}
\title{Print power and sample size results for paired mean difference}
\usage{
\method{print}{designPairedMeanDiff}(x, ...)
}
\arguments{
\item{x}{The designPairedMeanDiff object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
paired mean difference.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
