% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designPairedMeanDiffEquiv}
\alias{print.designPairedMeanDiffEquiv}
\title{Print power and sample size results for equivalence in paired
mean difference}
\usage{
\method{print}{designPairedMeanDiffEquiv}(x, ...)
}
\arguments{
\item{x}{The designPairedMeanDiffEquiv object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
equivalence in paired mean difference.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
