% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designPairedMeanRatio}
\alias{print.designPairedMeanRatio}
\title{Print power and sample size results for paired mean ratio}
\usage{
\method{print}{designPairedMeanRatio}(x, ...)
}
\arguments{
\item{x}{The designPairedMeanRatio object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
paired mean ratio.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
