% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designRepeatedANOVAContrast}
\alias{print.designRepeatedANOVAContrast}
\title{Print the power and sample size for a single contrast in
one-way repeated measures ANOVA}
\usage{
\method{print}{designRepeatedANOVAContrast}(x, ...)
}
\arguments{
\item{x}{The designRepeatedANOVAContrast object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the power and sample size for a single contrast in
one-way repeated measures analysis of variance.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
