% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designRiskDiff}
\alias{print.designRiskDiff}
\title{Print Power and Sample Size Results for Two-Sample Risk Difference}
\usage{
\method{print}{designRiskDiff}(x, ...)
}
\arguments{
\item{x}{The designRiskDiff object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
two-sample risk difference.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
