% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designRiskRatioFM}
\alias{print.designRiskRatioFM}
\title{Print power and sample size results for two-sample risk ratio
based on the Farrington-Manning score test}
\usage{
\method{print}{designRiskRatioFM}(x, ...)
}
\arguments{
\item{x}{The designRiskRatioFM object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
two-sample risk ratio based on the Farrington-Manning score test.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
