% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designSlopeDiff}
\alias{print.designSlopeDiff}
\title{Print power and sample size results for two-sample slope difference}
\usage{
\method{print}{designSlopeDiff}(x, ...)
}
\arguments{
\item{x}{The designSlopeDiff object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
two-sample slope difference.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
