% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designSlopeDiffMMRM}
\alias{print.designSlopeDiffMMRM}
\title{Print power and sample size results for two-sample slope difference
at the last time point from the MMRM model}
\usage{
\method{print}{designSlopeDiffMMRM}(x, ...)
}
\arguments{
\item{x}{The designSlopeDiffMMRM object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
two-sample slope difference at the last time point from the MMRM model.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
