% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designTwoMultinom}
\alias{print.designTwoMultinom}
\title{Print the power and sample size for difference in two-sample
multinomial response}
\usage{
\method{print}{designTwoMultinom}(x, ...)
}
\arguments{
\item{x}{The designTwoMultinom object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the power and sample size for difference in
two-sample multinomial response.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
