% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designTwoOrdinal}
\alias{print.designTwoOrdinal}
\title{Print the power and sample size for Wilcoxon test for two-sample
ordinal response.}
\usage{
\method{print}{designTwoOrdinal}(x, ...)
}
\arguments{
\item{x}{The designTwoOrdinal object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the power and sample size for Wilcoxon test for
two-sample ordinal response.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
