% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designTwoWayANOVA}
\alias{print.designTwoWayANOVA}
\title{Print the power and sample size for two-way ANOVA}
\usage{
\method{print}{designTwoWayANOVA}(x, ...)
}
\arguments{
\item{x}{The designTwoWayANOVA object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the power and sample size for two-way analysis
of variance.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
