% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designUnorderedBinom}
\alias{print.designUnorderedBinom}
\title{Print the power and sample size for unordered multi-sample
binomial response}
\usage{
\method{print}{designUnorderedBinom}(x, ...)
}
\arguments{
\item{x}{The designUnorderedBinom object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the power and sample size for unordered multi-sample
binomial response.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
