% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designUnorderedMultinom}
\alias{print.designUnorderedMultinom}
\title{Print the power and sample size for unordered multi-sample
multinomial response}
\usage{
\method{print}{designUnorderedMultinom}(x, ...)
}
\arguments{
\item{x}{The designUnorderedMultinom object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the power and sample size for unordered multi-sample
multinomial response.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
