% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.estimateCI}
\alias{print.estimateCI}
\title{Print the point estimate and confidence interval}
\usage{
\method{print}{estimateCI}(x, ...)
}
\arguments{
\item{x}{The estimateCI object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the point estimate and confidence interval.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
