% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.kmpower}
\alias{print.kmpower}
\title{Print Power and Sample Size Results for Milestone Survival
Difference}
\usage{
\method{print}{kmpower}(x, ...)
}
\arguments{
\item{x}{The kmpower object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from power
calculation.
}
\description{
Prints the summary statistics from power calculation.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
