% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.kmpower1s}
\alias{print.kmpower1s}
\title{Print power and sample size results for one-sample milestone
survival probability}
\usage{
\method{print}{kmpower1s}(x, ...)
}
\arguments{
\item{x}{The kmpower1s object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from power
calculation.
}
\description{
Prints the summary statistics from power calculation of
one-sample milestone survival probability.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
