% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.kmpowerequiv}
\alias{print.kmpowerequiv}
\title{Print Power and Sample Size Results for Equivalence in Milestone
Survival Probability Difference}
\usage{
\method{print}{kmpowerequiv}(x, ...)
}
\arguments{
\item{x}{The kmpowerequiv object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from power
calculation.
}
\description{
Prints the summary statistics from power calculation of
equivalence in milestone survival probability difference.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
