% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.lrpower}
\alias{print.lrpower}
\title{Print power and sample size results for log-rank tests}
\usage{
\method{print}{lrpower}(x, ...)
}
\arguments{
\item{x}{The lrpower object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from power
calculation.
}
\description{
Prints the summary statistics from power calculation.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
