% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.lrsim_bmTrtSel}
\alias{print.lrsim_bmTrtSel}
\title{Print Simulation Results for Phase 2/3 Seamless Design with
Biomarker-Based Treatment Selection}
\usage{
\method{print}{lrsim_bmTrtSel}(x, ...)
}
\arguments{
\item{x}{The lrsim_bmTrtSel object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from simulation runs.
}
\description{
Prints the summary statistics from simulation for a seamless
phase II/III design that selects one dose at the end of phase II based on
a short-term efficacy endpoint and a short-term toxicity endpoint, and
compares the testing procedures available for the confirmatory analysis.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
