% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.lrsim_mams}
\alias{print.lrsim_mams}
\title{Print Simulation Results for Multi-Arm Multi-Stage Design}
\usage{
\method{print}{lrsim_mams}(x, ...)
}
\arguments{
\item{x}{The lrsim_mams object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from simulation runs.
}
\description{
Prints the summary statistics from simulation.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
