% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.lrsim_mcpmod}
\alias{print.lrsim_mcpmod}
\title{Print Simulation Results for MCPMod Design for Log-Rank Test}
\usage{
\method{print}{lrsim_mcpmod}(x, ...)
}
\arguments{
\item{x}{The lrsim_mcpmod object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from simulation runs.
}
\description{
Prints the summary statistics from simulation for MCPMod design.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
