% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.lrsim_tsssd}
\alias{print.lrsim_tsssd}
\title{Print Simulation Results for Two-Stage Seamless Sequential Design}
\usage{
\method{print}{lrsim_tsssd}(x, ...)
}
\arguments{
\item{x}{The lrsim_tsssd object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from simulation runs.
}
\description{
Prints the summary statistics from simulation.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
