% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.mTPI2Table}
\alias{print.mTPI2Table}
\title{Print the decision table for the mTPI-2 design}
\usage{
\method{print}{mTPI2Table}(x, ...)
}
\arguments{
\item{x}{The mTPI2Table object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the decision table of the modified toxicity
probability-2 (mTPI-2) design for MTD finding.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
