% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.multiarm}
\alias{print.multiarm}
\title{Print Multi-Arm Multi-Stage Design}
\usage{
\method{print}{multiarm}(x, ...)
}
\arguments{
\item{x}{The design object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the stopping boundaries and power for
multi-arm multi-stage design.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
