% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.nbpower1s}
\alias{print.nbpower1s}
\title{Print Power and Sample Size Results for One-Sample Negative
Binomial Rate}
\usage{
\method{print}{nbpower1s}(x, ...)
}
\arguments{
\item{x}{The nbpower1s object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from power
calculation.
}
\description{
Prints the summary statistics from power calculation of
one-sample negative binomial rate.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
