% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.nbpowerequiv}
\alias{print.nbpowerequiv}
\title{Print Power and Sample Size Results for Equivalence in Negative
Binomial Rate Ratio}
\usage{
\method{print}{nbpowerequiv}(x, ...)
}
\arguments{
\item{x}{The nbpowerequiv object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from power
calculation.
}
\description{
Prints the summary statistics from power calculation of
equivalence in negative binomial rate ratio.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
