% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.rdsim_multiarm}
\alias{print.rdsim_multiarm}
\title{Print Simulation Results for Multi-Arm Multi-Stage Design for Risk
Difference}
\usage{
\method{print}{rdsim_multiarm}(x, ...)
}
\arguments{
\item{x}{The rdsim_multiarm object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from simulation runs.
}
\description{
Prints the summary statistics from risk-difference simulation.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
