% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.rmpower1s}
\alias{print.rmpower1s}
\title{Print Power and Sample Size Results for One-Sample Restricted
Mean Survival Time}
\usage{
\method{print}{rmpower1s}(x, ...)
}
\arguments{
\item{x}{The rmpower1s object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from power
calculation.
}
\description{
Prints the summary statistics from power calculation of
one-sample restricted mean survival time.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
