% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.rmpowerequiv}
\alias{print.rmpowerequiv}
\title{Print Power and Sample Size Results for Equivalence in Restricted
Mean Survival Time Difference}
\usage{
\method{print}{rmpowerequiv}(x, ...)
}
\arguments{
\item{x}{The rmpowerequiv object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from power
calculation.
}
\description{
Prints the summary statistics from power calculation of
equivalence in restricted mean survival time difference.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
