% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.tsssd}
\alias{print.tsssd}
\title{Print Two-Stage Seamless Sequential Design}
\usage{
\method{print}{tsssd}(x, ...)
}
\arguments{
\item{x}{The design object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the stopping boundaries and power for
two-stage seamless sequential design.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
