% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Run Shiny app}
\usage{
runShinyApp()
}
\description{
Runs the log-rank test power and sample size calculation
Shiny app.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
