% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp_lrstat.R
\name{runShinyApp_lrstat}
\alias{runShinyApp_lrstat}
\title{Run Shiny App}
\usage{
runShinyApp_lrstat()
}
\value{
No return value, called for side effects.
}
\description{
Runs the log-rank test power and sample size calculation
Shiny app.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
