% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_seed}
\alias{set_seed}
\title{Set seed}
\usage{
set_seed(seed)
}
\arguments{
\item{seed}{The seed to use for generating random numbers.}
}
\value{
No return value, called for side effects.
}
\description{
Sets the R seed in the cpp program based on set.seed() in R.
}
\examples{
set_seed(123)

}
\keyword{internal}
