% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stl_sort}
\alias{stl_sort}
\title{Sort a numeric vector}
\usage{
stl_sort(x)
}
\arguments{
\item{x}{The numeric vector to sort.}
}
\value{
A vector obtained after sorting the input vector.
}
\description{
Sorts a numeric vector in the cpp program.
}
\examples{
stl_sort(c(3, 4.2, 1))

}
\keyword{internal}
