% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zstatRateDiff}
\alias{zstatRateDiff}
\title{Miettinen-Nurminen Score Test Statistic for Two-Sample Rate
Difference}
\usage{
zstatRateDiff(t1, y1, t2, y2, rateDiffH0 = 0)
}
\arguments{
\item{t1}{The exposure for the active treatment group.}

\item{y1}{The number of events for the active treatment group.}

\item{t2}{The exposure for the control group.}

\item{y2}{The number of events for the control group.}

\item{rateDiffH0}{The rate difference under the null hypothesis.
Defaults to 0.}
}
\value{
The value of the score test statistic.
}
\description{
Obtains the Miettinen-Nurminen score test statistic for
two-sample rate difference possibly with stratification.
}
\details{
The Mantel-Haenszel weights are used for stratified samples.
}
\examples{

zstatRateDiff(t1 = c(10, 10), y1 = c(4, 3),
              t2 = c(20, 10), y2 = c(2, 0), rateDiffH0 = 0)

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
