% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adaptDesign}
\alias{adaptDesign}
\title{Adaptive design at an interim look}
\usage{
adaptDesign(
  beta = NA_real_,
  INew = NA_real_,
  L = NA_integer_,
  zL = NA_real_,
  theta = NA_real_,
  kMax = NA_integer_,
  informationRates = NA_real_,
  criticalValues = NA_real_,
  futilityBounds = NA_real_,
  MullerSchafer = 0L,
  kNew = NA_integer_,
  tNew = NA_real_,
  efficacyStopping = NA_integer_,
  futilityStopping = NA_integer_,
  typeAlphaSpending = "sfOF",
  parameterAlphaSpending = NA_real_,
  typeBetaSpending = "none",
  parameterBetaSpending = NA_real_,
  userBetaSpending = NA_real_,
  spendingTime = NA_real_
)
}
\arguments{
\item{beta}{The conditional type II error.}

\item{INew}{The incremental information.}

\item{L}{The interim look.}

\item{zL}{The Z test statistic at the interim look.}

\item{theta}{The parameter value.}

\item{kMax}{The maximum number of stages of the original design.}

\item{informationRates}{The information rates of the original design.}

\item{criticalValues}{The upper boundaries on the Z-test statistic scale
for efficacy stopping for the original design.}

\item{futilityBounds}{The lower boundaries on the Z-test statistic scale
for futility stopping for the original design.}

\item{MullerSchafer}{Whether to use the Muller and Schafer (2001) method
for trial adaptation.}

\item{kNew}{The number of future looks.}

\item{tNew}{The spacing of future looks in terms of information rates.}

\item{efficacyStopping}{The indicators of whether efficacy stopping is
allowed at each future look. Defaults to true if left unspecified.}

\item{futilityStopping}{The indicators of whether futility stopping is
allowed at each future look. Defaults to true if left unspecified.}

\item{typeAlphaSpending}{The type of alpha spending for future looks.
One of the following:
"OF" for O'Brien-Fleming boundaries,
"P" for Pocock boundaries,
"WT" for Wang & Tsiatis boundaries,
"sfOF" for O'Brien-Fleming type spending function,
"sfP" for Pocock type spending function,
"sfKD" for Kim & DeMets spending function,
"sfHSD" for Hwang, Shi & DeCani spending function, and
"none" for no early efficacy stopping.
Defaults to "sfOF".}

\item{parameterAlphaSpending}{The parameter value for the alpha
spending for future looks.
Corresponds to Delta for "WT", rho for "sfKD", and gamma for "sfHSD".}

\item{typeBetaSpending}{The type of beta spending for future looks.
One of the following:
"sfOF" for O'Brien-Fleming type spending function,
"sfP" for Pocock type spending function,
"sfKD" for Kim & DeMets spending function,
"sfHSD" for Hwang, Shi & DeCani spending function,
"user" for user defined spending, and
"none" for no early futility stopping.
Defaults to "none".}

\item{parameterBetaSpending}{The parameter value for the beta spending
for future looks.
Corresponds to rho for "sfKD", and gamma for "sfHSD".}

\item{userBetaSpending}{The user defined cumulative beta spent at
future looks.}

\item{spendingTime}{A vector of length \code{kNew} for the error spending
time at future looks. Defaults to missing, in which case, it is the
same as \code{tNew}.}
}
\value{
An \code{adaptDesign} object with two list components:
\itemize{
\item \code{primaryTrial}: A list of selected information for the primary
trial, including \code{L}, \code{zL}, \code{theta}, \code{kMax},
\code{informationRates}, \code{efficacyBounds}, \code{futilityBounds},
and \code{MullerSchafer}.
\item \code{secondaryTrial}: A \code{design} object for the secondary trial.
}
}
\description{
Obtains the conditional power for specified incremental
information given the interim results, parameter value, and data-dependent
changes in the error spending function and the number and spacing of
interim looks. Conversely, obtain the incremental information needed
to attain a specified conditional power given the interim results,
parameter value, and data-dependent changes in the error spending
function and the number and spacing of interim looks.
}
\examples{

# original group sequential design with 90\% power to detect delta = 6
delta = 6
sigma = 17
n = 282
(des1 = getDesign(Imax = n/(4*sigma^2), theta = delta, kMax = 3, 
                  alpha = 0.05, typeAlphaSpending = "sfHSD", 
                  parameterAlphaSpending = -4))

# interim look results
L = 1
n1 = n/3
delta1 = 4.5
sigma1 = 20
zL = delta1/sqrt(4/n1*sigma1^2)

# conditional power for original design at estimated parameter value
(des2 = adaptDesign(
  beta = NA, INew = (n-n1)/(4*sigma1^2), L, zL, theta = delta1, 
  kMax = des1$overallResults$kMax, 
  informationRates = des1$byStageResults$informationRates,
  criticalValues = des1$byStageResults$efficacyBounds))

# conditional power with sample size increase
(des2 = adaptDesign(
  beta = NA, INew = 420/(4*sigma1^2), L, zL, theta = delta1, 
  kMax = des1$overallResults$kMax, 
  informationRates = des1$byStageResults$informationRates,
  criticalValues = des1$byStageResults$efficacyBounds))

# Muller & Schafer (2001) method to design the secondary trial: 
# 3-look gamma(-2) spending with equal spacing, 84\% power at delta = 4.5
(des2 = adaptDesign(
  beta = 0.16, INew = NA, L, zL, theta = delta1,
  kMax = des1$overallResults$kMax, 
  informationRates = des1$byStageResults$informationRates,
  criticalValues = des1$byStageResults$efficacyBounds,
  MullerSchafer = TRUE,
  kNew = 3, typeAlphaSpending = "sfHSD", 
  parameterAlphaSpending = -2))
  
# incremental sample size for sigma = 20
(nNew = 4*sigma1^2*des2$overallResults$maxInformation)

}
\seealso{
\code{\link{getDesign}}
}
