% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{finthyp}
\alias{finthyp}
\title{Indicator Matrix for All Intersection Hypotheses}
\usage{
finthyp(m)
}
\arguments{
\item{m}{The number of elementary hypotheses.}
}
\value{
The \eqn{(2^m-1) \times m} indicator matrix for the intersection
hypotheses, with row 1 the global intersection (all hypotheses active)
and the last row the singleton \code{{m}}.
}
\description{
Obtains the indicator matrix for all intersection hypotheses
formed from a given number of elementary hypotheses, matching the row
order used by \code{fwgtmat} and \code{fDefaultWgtmat}.
}
\examples{

finthyp(3)

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
