% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_two_stage.R
\name{fStageBound}
\alias{fStageBound}
\title{Compute stage-wise bounds for multiple testing}
\usage{
fStageBound(
  m,
  wgtmat,
  family = NULL,
  corr = NULL,
  alpha,
  alpha1,
  info_frac,
  nthreads = 0
)
}
\arguments{
\item{m}{Number of hypotheses}

\item{wgtmat}{Weight matrix for the intersection hypotheses}

\item{family}{Family matrix indicating which hypotheses belong to which
families. The correlation is known only for hypotheses belonging to the
same family. If NULL, all hypotheses are assumed to belong to the same
family.}

\item{corr}{Correlation matrix for the test statistics}

\item{alpha}{Overall significance level}

\item{alpha1}{Significance level for the first stage}

\item{info_frac}{Information fraction for the first stage}

\item{nthreads}{The number of threads to use in simulations (0 means
the default RcppParallel behavior).}
}
\value{
A list containing the stage 1 and stage 2 bounds for the individual
hypotheses in each intersection hypothesis.
}
\description{
Compute stage-wise bounds for a two-stage multiple testing
procedure.
}
\examples{
initial_weights <- c(0.5, 0.5, 0, 0)
transition_matrix <- matrix(c(0, 0.5, 0.5, 0,
                              0.5, 0, 0, 0.5,
                              0, 1, 0, 0,
                              1, 0, 0, 0),
                            nrow = 4, byrow = TRUE)
wgtmat <- fwgtmat(initial_weights, transition_matrix)
family <- matrix(c(1, 1, 0, 0,
                   0, 0, 1, 1),
                 nrow = 2, byrow = TRUE)
corr <- matrix(c(1, 0.5, NA, NA,
                 0.5, 1, NA, NA,
                 NA, NA, 1, 0.5,
                 NA, NA, 0.5, 1),
               nrow = 4, byrow = TRUE)
alpha <- 0.025
alpha1 <- errorSpent(0.5, alpha, "sfOF")
fStageBound(4, wgtmat, family, corr, alpha, alpha1, 0.5,
            nthreads = 1)

}
\references{
Cyrus Mehta, Ajoy Mukhopadhyay, and Martin Posch. Graph Based, Adaptive,
Multiarm, Multiple Endpoint, Two-Stage Designs. Statistics in Medicine.
2025.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
