% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getADCI_seamless}
\alias{getADCI_seamless}
\title{Confidence Interval After Adaptation for a Phase 2/3 Seamless
Design}
\usage{
getADCI_seamless(
  M = NA_integer_,
  r = 1,
  corr_known = TRUE,
  L = NA_integer_,
  zL = NA_real_,
  IMax = NA_real_,
  K = NA_integer_,
  informationRates = NA_real_,
  efficacyStopping = NA_integer_,
  criticalValues = NA_real_,
  alpha = 0.25,
  typeAlphaSpending = "sfOF",
  parameterAlphaSpending = NA_real_,
  spendingTime = NA_real_,
  MullerSchafer = FALSE,
  Lc = NA_integer_,
  zLc = NA_real_,
  INew = NA_real_,
  informationRatesNew = NA_real_,
  efficacyStoppingNew = NA_integer_,
  typeAlphaSpendingNew = "sfOF",
  parameterAlphaSpendingNew = NA_real_,
  spendingTimeNew = NA_real_
)
}
\arguments{
\item{M}{Number of active treatment arms in Phase 2.}

\item{r}{Randomization ratio of each active arm to the common control
in Phase 2.}

\item{corr_known}{Logical. If \code{TRUE}, the correlation between Wald
statistics in Phase 2 is derived from the randomization ratio \eqn{r}
as \eqn{r / (r + 1)}. If \code{FALSE}, a conservative correlation of
0 is used.}

\item{L}{The interim adaptation look in Phase 3.}

\item{zL}{The z-test statistic at the interim adaptation look of
Phase 3.}

\item{IMax}{Maximum information for the active arm versus the common
control for the original trial. Must be provided.}

\item{K}{Number of sequential looks in Phase 3.}

\item{informationRates}{A numeric vector of information rates fixed
before the trial. If unspecified, defaults to \eqn{(1:(K+1)) / (K+1)}.}

\item{efficacyStopping}{Indicators of whether efficacy stopping is
allowed at each stage of the primary trial. Defaults to \code{TRUE}
if left unspecified.}

\item{criticalValues}{The upper boundaries on the max z-test statistic
scale for Phase 2 and the z-test statistics for the selected arm
in Phase 3 for the primary trial. If missing, boundaries
will be computed based on the specified alpha spending function.}

\item{alpha}{The significance level of the primary trial.
Defaults to 0.025.}

\item{typeAlphaSpending}{The type of alpha spending for the primary
trial. One of the following:
\code{"OF"} for O'Brien-Fleming boundaries,
\code{"P"} for Pocock boundaries,
\code{"WT"} for Wang & Tsiatis boundaries,
\code{"sfOF"} for O'Brien-Fleming type spending function,
\code{"sfP"} for Pocock type spending function,
\code{"sfKD"} for Kim & DeMets spending function,
\code{"sfHSD"} for Hwang, Shi & DeCani spending function, and
\code{"none"} for no early efficacy stopping.
Defaults to \code{"sfOF"}.}

\item{parameterAlphaSpending}{The parameter value of alpha spending
for the primary trial. Corresponds to \eqn{\Delta} for \code{"WT"},
\eqn{\rho} for \code{"sfKD"}, and \eqn{\gamma} for \code{"sfHSD"}.}

\item{spendingTime}{The error spending time of the primary trial.
Defaults to missing, in which case, it is the same as
\code{informationRates}.}

\item{MullerSchafer}{Whether to use the Muller and Schafer (2001) method
for trial adaptation.}

\item{Lc}{The termination look of the integrated trial.}

\item{zLc}{The z-test statistic at the termination look of the
integrated trial.}

\item{INew}{The maximum information for the active arm versus the common
control in the secondary trial.}

\item{informationRatesNew}{The spacing of looks of the secondary trial.}

\item{efficacyStoppingNew}{The indicators of whether efficacy stopping is
allowed at each look of the secondary trial.
Defaults to \code{TRUE} if left unspecified.}

\item{typeAlphaSpendingNew}{The type of alpha spending for the secondary
trial. One of the following:
\code{"OF"} for O'Brien-Fleming boundaries,
\code{"P"} for Pocock boundaries,
\code{"WT"} for Wang & Tsiatis boundaries,
\code{"sfOF"} for O'Brien-Fleming type spending function,
\code{"sfP"} for Pocock type spending function,
\code{"sfKD"} for Kim & DeMets spending function,
\code{"sfHSD"} for Hwang, Shi & DeCani spending function, and
\code{"none"} for no early efficacy stopping.
Defaults to \code{"sfOF"}.}

\item{parameterAlphaSpendingNew}{The parameter value of alpha spending
for the secondary trial. Corresponds to \eqn{\Delta} for \code{"WT"},
\eqn{\rho} for \code{"sfKD"}, and \eqn{\gamma} for \code{"sfHSD"}.}

\item{spendingTimeNew}{The error spending time of the secondary trial.
Defaults to missing, in which case, it is
the same as \code{informationRatesNew}.}
}
\value{
A data frame with the following variables:
\itemize{
\item \code{pvalue}: p-value for rejecting the null hypothesis.
\item \code{thetahat}: Point estimate of the parameter.
\item \code{cilevel}: Confidence interval level.
\item \code{lower}: Lower bound of confidence interval.
\item \code{upper}: Upper bound of confidence interval.
}
}
\description{
Obtains the p-value, conservative point estimate, and
confidence interval after the end of an adaptive phase 2/3 seamless
design.
}
\details{
If typeAlphaSpendingNew is \code{"OF"}, \code{"P"}, \code{"WT"},
or \code{"none"}, then
\code{informationRatesNew}, \code{efficacyStoppingNew}, and
\code{spendingTimeNew} must be of full length \code{kNew}, and
\code{informationRatesNew} and \code{spendingTimeNew} must end with 1.
}
\examples{
getADCI_seamless(
  M = 2, r = 1, corr_known = FALSE,
  L = 1, zL = -log(0.67) * sqrt(80 / 4),
  IMax = 120 / 4, K = 2, informationRates = c(1/3, 2/3, 1),
  alpha = 0.025, typeAlphaSpending = "OF",
  Lc = 2, zLc = -log(0.677) * sqrt(236 / 4), INew = 236 / 4)

}
\references{
Ping Gao, Yingqiu Li.
Adaptive multiple comparison sequential design (AMCSD) for clinical trials.
Journal of Biopharmaceutical Statistics, 2024, 34(3), 424-440.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
