% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDesignProportions.R
\name{getDesignOneMultinom}
\alias{getDesignOneMultinom}
\title{Power and sample for one-sample multinomial response}
\usage{
getDesignOneMultinom(
  beta = NA_real_,
  n = NA_real_,
  ncats = NA_integer_,
  piH0 = NA_real_,
  pi = NA_real_,
  rounding = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{beta}{The type II error.}

\item{n}{The total sample size.}

\item{ncats}{The number of categories of the multinomial response.}

\item{piH0}{The prevalence of each category under the null hypothesis.
Only need to provide the values for the first \code{ncats-1}
categories.}

\item{pi}{The prevalence of each category. Only need to provide the
values for the first \code{ncats-1} categories.}

\item{rounding}{Whether to round up sample size. Defaults to 1 for
sample size rounding.}

\item{alpha}{The two-sided significance level. Defaults to 0.05.}
}
\value{
An S3 class \code{designOneMultinom} object with the following
components:
\itemize{
\item \code{power}: The power to reject the null hypothesis.
\item \code{alpha}: The two-sided significance level.
\item \code{n}: The maximum number of subjects.
\item \code{ncats}: The number of categories of the multinomial response.
\item \code{piH0}: The prevalence of each category under the null hypothesis.
\item \code{pi}: The prevalence of each category.
\item \code{effectsize}: The effect size for the chi-square test.
\item \code{calculationTarget}: The calculation target, \code{beta} or
\code{n}.
\item \code{rounding}: Whether to round up sample size.
}
}
\description{
Obtains the power given sample size or obtains the sample
size given power for one-sample multinomial response.
}
\examples{

(design1 <- getDesignOneMultinom(
  beta = 0.1, ncats = 3, piH0 = c(0.25, 0.25),
  pi = c(0.3, 0.4), alpha = 0.05))

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
