% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_futilityBounds}
\alias{param_futilityBounds}
\title{Parameter Description: futilityBounds}
\arguments{
\item{futilityBounds}{Lower boundaries on the z-test statistic scale
for stopping for futility at stages 1, ..., \code{kMax-1}. Defaults to
\code{rep(-6, kMax-1)} if left unspecified.}
}
\description{
Parameter Description: futilityBounds
}
\keyword{internal}
