% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_lambda1_stratified}
\alias{param_lambda1_stratified}
\title{Parameter Description: lambda1 stratified}
\arguments{
\item{lambda1}{A vector of hazard rates for the event for the
active treatment group, one for each analysis time interval, by stratum.}
}
\description{
Parameter Description: lambda1 stratified
}
\keyword{internal}
