% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_parameterAlphaSpending}
\alias{param_parameterAlphaSpending}
\title{Parameter Description: parameterAlphaSpending}
\arguments{
\item{parameterAlphaSpending}{The parameter value for the alpha spending.
Corresponds to Delta for "WT", rho for "sfKD", and gamma for "sfHSD".}
}
\description{
Parameter Description: parameterAlphaSpending
}
\keyword{internal}
