% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_typeBetaSpending}
\alias{param_typeBetaSpending}
\title{Parameter Description: typeBetaSpending}
\arguments{
\item{typeBetaSpending}{The type of beta spending. One of the following:
"sfOF" for O'Brien-Fleming type spending function, "sfP" for Pocock type
spending function, "sfKD" for Kim & DeMets spending function,
"sfHSD" for Hwang, Shi & DeCani spending function, "user" for
user defined spending, and "none" for no early futility stopping.
Defaults to "none".}
}
\description{
Parameter Description: typeBetaSpending
}
\keyword{internal}
