% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_descriptions.R
\name{param_typeOfComputation}
\alias{param_typeOfComputation}
\title{Parameter Description: typeOfComputation}
\arguments{
\item{typeOfComputation}{The type of computation,
either "direct" for the direct approximation method,
or "schoenfeld" for the Schoenfeld method.
Defaults to "direct". Can use "Schoenfeld"
under proportional hazards and conventional log-rank test.}
}
\description{
Parameter Description: typeOfComputation
}
\keyword{internal}
