% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormr.R
\name{pmvnormr}
\alias{pmvnormr}
\title{Probability of Multivariate Normal Distribution over a Hyperrectangle}
\usage{
pmvnormr(lower, upper, mean = NULL, sigma, ..., nthreads = 0)
}
\arguments{
\item{lower}{the lower bounds of the hyperrectangle.}

\item{upper}{the upper bounds of the hyperrectangle.}

\item{mean}{the mean vector (optional). If NULL, defaults to rep(0, nrow(sigma)).}

\item{sigma}{the covariance matrix of the multivariate normal distribution.}

\item{...}{forwarded to pmvnormRcpp (fast, n0, n_max, R, abseps, releps,
seed, parallel)}

\item{nthreads}{the number of threads to use in simulations (0 means
the default RcppParallel behavior).}
}
\value{
list(prob=..., error=..., n=...)
}
\description{
Computes the probability that a multivariate normal random vector
falls within a specified hyperrectangle defined by lower and upper bounds.
}
\examples{
n <- 5
mean <- rep(0, 5)
lower <- rep(-1, 5)
upper <- rep(3, 5)
corr <- diag(5)
corr[lower.tri(corr)] <- 0.5
corr[upper.tri(corr)] <- 0.5
prob <- pmvnormr(lower, upper, mean, corr)
print(prob)

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
