% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designMeanDiffCarryover}
\alias{print.designMeanDiffCarryover}
\title{Print power and sample size results for direct treatment effects
in crossover trials accounting for carryover effects}
\usage{
\method{print}{designMeanDiffCarryover}(x, ...)
}
\arguments{
\item{x}{The designMeanDiffCarryover object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
direct treatment effects in crossover trials accounting for
carryover effects.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
