% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designMeanDiffMMRM}
\alias{print.designMeanDiffMMRM}
\title{Print Power and Sample Size Results for Two-Sample Mean Difference
at the Last Time Point From the MMRM Model}
\usage{
\method{print}{designMeanDiffMMRM}(x, ...)
}
\arguments{
\item{x}{The designMeanDiffMMRM object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
two-sample mean difference at the last time point from the MMRM model.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
