% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designMeanDiffXOEquiv}
\alias{print.designMeanDiffXOEquiv}
\title{Print Power and Sample Size Results for Equivalence in
Mean Difference in 2x2 Crossover}
\usage{
\method{print}{designMeanDiffXOEquiv}(x, ...)
}
\arguments{
\item{x}{The designMeanDiffXOEquiv object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
equivalence in mean difference in 2x2 crossover.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
