% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designMeanRatioEquiv}
\alias{print.designMeanRatioEquiv}
\title{Print Power and Sample Size Results for Equivalence in
Two-Sample Mean Ratio}
\usage{
\method{print}{designMeanRatioEquiv}(x, ...)
}
\arguments{
\item{x}{The designMeanRatioEquiv object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
equivalence in two-sample mean ratio.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
