% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designPairedPropMcNemar}
\alias{print.designPairedPropMcNemar}
\title{Print power and sample size results for McNemar's test for paired
proportions}
\usage{
\method{print}{designPairedPropMcNemar}(x, ...)
}
\arguments{
\item{x}{The designPairedPropMcNemar object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
McNemar's test for paired proportions.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
