% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.designWilcoxon}
\alias{print.designWilcoxon}
\title{Print power and sample size results for two-sample Wilcoxon test}
\usage{
\method{print}{designWilcoxon}(x, ...)
}
\arguments{
\item{x}{The designWilcoxon object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the summary statistics from power calculation of
two-sample Wilcoxon test.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
