% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.lrpowerequiv}
\alias{print.lrpowerequiv}
\title{Print Power and Sample Size Results for Equivalence in Hazard
Ratio}
\usage{
\method{print}{lrpowerequiv}(x, ...)
}
\arguments{
\item{x}{The lrpowerequiv object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from power
calculation.
}
\description{
Prints the summary statistics from power calculation of
equivalence in hazard ratio.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
