% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.lrsim}
\alias{print.lrsim}
\title{Print simulation results for log-rank tests}
\usage{
\method{print}{lrsim}(x, ...)
}
\arguments{
\item{x}{The lrsim object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from simulation runs.
}
\description{
Prints the summary statistics from simulation.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
