% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.lrsim_multiarm}
\alias{print.lrsim_multiarm}
\title{Print Simulation Results for Multi-Arm Multi-Stage Design for Logrank Test}
\usage{
\method{print}{lrsim_multiarm}(x, ...)
}
\arguments{
\item{x}{The lrsim_multiarm object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from simulation runs.
}
\description{
Prints the summary statistics from simulation for logrank test.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
