% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.lrsim_seamless}
\alias{print.lrsim_seamless}
\title{Print Simulation Results for Phase 2/3 Seamless Design}
\usage{
\method{print}{lrsim_seamless}(x, ...)
}
\arguments{
\item{x}{The lrsim_seamless object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from simulation runs.
}
\description{
Prints the summary statistics from simulation.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
