% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.nbpower}
\alias{print.nbpower}
\title{Print power and sample size results for negative binomial rate
ratio}
\usage{
\method{print}{nbpower}(x, ...)
}
\arguments{
\item{x}{The nbpower object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the summary statistics from power
calculation.
}
\description{
Prints the summary statistics from power calculation of
negative binomial rate ratio.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
