% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.seamless}
\alias{print.seamless}
\title{Print Phase 2/3 Seamless Design}
\usage{
\method{print}{seamless}(x, ...)
}
\arguments{
\item{x}{The design object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A tabular printout of the design elements.
}
\description{
Prints the stopping boundaries and power for
phase 2/3 seamless design.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
