% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormr.R
\name{qmvnormr}
\alias{qmvnormr}
\title{Quantile of the Multivariate Normal Distribution}
\usage{
qmvnormr(
  p,
  mean = NULL,
  sigma,
  fast = TRUE,
  n0 = 128,
  n_max = 16384,
  R = 8,
  abseps = 1e-04,
  releps = 0,
  seed = 0,
  parallel = TRUE,
  nthreads = 0
)
}
\arguments{
\item{p}{the probability.}

\item{mean}{the mean vector (optional). If NULL, defaults to
\code{rep(0, nrow(sigma))}.}

\item{sigma}{the covariance matrix of the multivariate normal distribution.}

\item{fast}{whether to use the fast version of pnorm and qnorm.}

\item{n0}{the initial number of samples for the Monte Carlo estimation
per replication.}

\item{n_max}{the maximum number of samples for the Monte Carlo estimation
per replication.}

\item{R}{the number of replications for the Monte Carlo estimation.}

\item{abseps}{the absolute error tolerance for the Monte Carlo estimation.}

\item{releps}{the relative error tolerance for the Monte Carlo estimation.}

\item{seed}{the random seed for reproducibility. If 0 (default), seed
will be generated randomly.}

\item{parallel}{whether to use parallel computation for the Monte Carlo
estimation.}

\item{nthreads}{the number of threads to use in simulations (0 means
the default RcppParallel behavior).}
}
\value{
The calculated quantile.
}
\description{
Computes the equicoordinate quantile function of the
multivariate normal distribution.
}
\examples{
n <- 5
mean <- rep(0, 5)
corr <- diag(5)
corr[lower.tri(corr)] <- 0.5
corr[upper.tri(corr)] <- 0.5
qmvnormr(0.5, mean, corr)

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
