% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormr.R
\name{qmvnormr}
\alias{qmvnormr}
\title{Quantile of the Multivariate Normal Distribution}
\usage{
qmvnormr(p, mean = NULL, sigma, ..., nthreads = 0)
}
\arguments{
\item{p}{the probability.}

\item{mean}{the mean vector (optional). If NULL, defaults to rep(0, nrow(sigma)).}

\item{sigma}{the covariance matrix of the multivariate normal distribution.}

\item{...}{forwarded to qmvnormRcpp (fast, n0, n_max, R, abseps, releps, seed)}

\item{nthreads}{the number of threads to use in simulations (0 means
the default RcppParallel behavior).}
}
\value{
The calculated quantile.
}
\description{
Computes the equicoordinate quantile function of the
multivariate normal distribution.
}
\examples{
n <- 5
mean <- rep(0, 5)
corr <- diag(5)
corr[lower.tri(corr)] <- 0.5
corr[upper.tri(corr)] <- 0.5
qmvnormr(0.5, mean, corr)

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
