% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remlRateDiff}
\alias{remlRateDiff}
\title{REML estimates of individual rates with specified rate
difference}
\usage{
remlRateDiff(rateDiffH0, t1, y1, t2, y2)
}
\arguments{
\item{rateDiffH0}{The specified rate difference.}

\item{t1}{The exposure for the active treatment group.}

\item{y1}{The number of events for the active treatment group.}

\item{t2}{The exposure for the control group.}

\item{y2}{The number of events for the control group.}
}
\value{
A vector of the restricted maximum likelihood estimates
of the incidence rates for the two treatment groups.
}
\description{
Obtains the restricted maximum likelihood estimates of
individual proportions with specified rate difference.
}
\examples{

remlRateDiff(rateDiffH0 = 0.1, t1 = 10, y1 = 4, t2 = 20, y2 = 2)

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
